Enum added in a separate file
